+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend.go
@@ -8,7 +8,9 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/kubernetes_annotation_value"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/kubernetes_label_key"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/kubernetes_label_value"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/api_container"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/label_key_consts"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/label_value_consts"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/object_name_constants"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/exec_result"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact_expander"
@@ -18,9 +20,48 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/wait_for_availability_http_methods"
+	"github.com/kurtosis-tech/stacktrace"
 	"io"
+	apiv1 "k8s.io/api/core/v1"
 	"net"
+	"strconv"
 )
+
+const (
+	// The Kurtosis servers (Engine and API Container) uses gRPC so MUST listen on TCP (no other protocols are supported), which also
+	// means that its grpc-proxy must listen on TCP
+	kurtosisServersPortProtocol = port_spec.PortProtocol_TCP
+
+	// The ID of the GRPC port for Kurtosis-internal containers (e.g. API container, engine, modules, etc.) which will
+	//  be stored in the port spec label
+	kurtosisInternalContainerGrpcPortSpecId = "grpc"
+	// The GRPC port protocol for Kurtosis-internal containers
+	kurtosisInternalContainerGrpcPortProtocol = apiv1.ProtocolTCP
+
+	// The ID of the GRPC proxy port for Kurtosis-internal containers. This is necessary because
+	// Typescript's grpc-web cannot communicate directly with GRPC ports, so Kurtosis-internal containers
+	// need a proxy  that will translate grpc-web requests before they hit the main GRPC server
+	kurtosisInternalContainerGrpcProxyPortSpecId = "grpcProxy"
+	// The GRPC proxy port protocol for Kurtosis-internal containers
+	kurtosisInternalContainerGrpcProxyPortProtocol = apiv1.ProtocolTCP
+
+	// Port number string parsing constants
+	publicPortNumStrParsingBase = 10
+	publicPortNumStrParsingBits = 16
+
+	externalServiceType = "ClusterIP"
+
+	sentencesSeparator = ", "
+)
+
+// This maps a Kubernetes pod's phase to a binary "is the pod considered running?" determiner
+// Its completeness is enforced via unit test
+var isPodRunningDeterminer = map[apiv1.PodPhase]bool{
+	apiv1.PodPending: true,
+	apiv1.PodRunning: true,
+	apiv1.PodSucceeded: false,
+	apiv1.PodUnknown: true, //We can not say that a pod is not running if we don't know the real state
+}
 
 type KubernetesKurtosisBackend struct {
 	kubernetesManager *kubernetes_manager.KubernetesManager
@@ -41,35 +82,6 @@ type KubernetesKurtosisBackend struct {
 }
 
 func (backend *KubernetesKurtosisBackend) PullImage(image string) error {
-	//TODO implement me
-	panic("implement me")
-}
-
-func (backend *KubernetesKurtosisBackend) CreateAPIContainer(
-	ctx context.Context,
-	image string,
-	enclaveId enclave.EnclaveID,
-	ipAddr net.IP,
-	grpcPortNum uint16,
-	grpcProxyPortNum uint16,
-	enclaveDataVolumeDirpath string,
-	envVars map[string]string,
-) (*api_container.APIContainer, error) {
-	//TODO implement me
-	panic("implement me")
-}
-
-func (backend *KubernetesKurtosisBackend) GetAPIContainers(ctx context.Context, filters *api_container.APIContainerFilters) (map[enclave.EnclaveID]*api_container.APIContainer, error) {
-	//TODO implement me
-	panic("implement me")
-}
-
-func (backend *KubernetesKurtosisBackend) StopAPIContainers(ctx context.Context, filters *api_container.APIContainerFilters) (successfulApiContainerIds map[enclave.EnclaveID]bool, erroredApiContainerIds map[enclave.EnclaveID]error, resultErr error) {
-	//TODO implement me
-	panic("implement me")
-}
-
-func (backend *KubernetesKurtosisBackend) DestroyAPIContainers(ctx context.Context, filters *api_container.APIContainerFilters) (successfulApiContainerIds map[enclave.EnclaveID]bool, erroredApiContainerIds map[enclave.EnclaveID]error, resultErr error) {
 	//TODO implement me
 	panic("implement me")
 }
@@ -227,4 +239,127 @@ func getStringMapFromAnnotationMap(labelMap map[*kubernetes_annotation_key.Kuber
 		strMap[labelKey.GetString()] = labelValue.GetString()
 	}
 	return strMap
+}
+
+// getPublicPortSpecFromServicePort returns a port_spec representing a kurtosis port spec for a service port in Kubernetes
+func getPublicPortSpecFromServicePort(servicePort apiv1.ServicePort, portProtocol port_spec.PortProtocol) (*port_spec.PortSpec, error) {
+	publicPortNumStr := strconv.FormatInt(int64(servicePort.Port), publicPortNumStrParsingBase)
+	publicPortNumUint64, err := strconv.ParseUint(publicPortNumStr, publicPortNumStrParsingBase, publicPortNumStrParsingBits)
+	if err != nil {
+		return nil, stacktrace.Propagate(
+			err,
+			"An error occurred parsing public port string '%v' using base '%v' and uint bits '%v'",
+			publicPortNumStr,
+			publicPortNumStrParsingBase,
+			publicPortNumStrParsingBits,
+		)
+	}
+	publicPortNum := uint16(publicPortNumUint64) // Safe to do because we pass the requisite number of bits into the parse command
+	publicGrpcPort, err := port_spec.NewPortSpec(publicPortNum, portProtocol)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "Expected to be able to create a port spec describing a public port on a Kubernetes node using number '%v' and protocol '%v', instead a non nil error was returned", publicPortNum, portProtocol)
+	}
+
+	return publicGrpcPort, nil
+}
+
+func getGrpcAndGrpcProxyPortSpecsFromServicePorts(servicePorts []apiv1.ServicePort) (resultGrpcPortSpec *port_spec.PortSpec, resultGrpcProxyPortSpec *port_spec.PortSpec, resultErr error) {
+	var publicGrpcPort *port_spec.PortSpec
+	var publicGrpcProxyPort *port_spec.PortSpec
+	grpcPortName := object_name_constants.KurtosisInternalContainerGrpcPortName.GetString()
+	grpcProxyPortName := object_name_constants.KurtosisInternalContainerGrpcProxyPortName.GetString()
+
+	for _, servicePort := range servicePorts {
+		servicePortName := servicePort.Name
+		switch servicePortName {
+		case grpcPortName:
+			{
+				publicGrpcPortSpec, err := getPublicPortSpecFromServicePort(servicePort, kurtosisServersPortProtocol)
+				if err != nil {
+					return nil, nil, stacktrace.Propagate(err, "Expected to be able to create a port spec describing a public grpc port from Kubernetes service port '%v', instead a non nil error was returned", servicePortName)
+				}
+				publicGrpcPort = publicGrpcPortSpec
+			}
+		case grpcProxyPortName:
+			{
+				publicGrpcProxyPortSpec, err := getPublicPortSpecFromServicePort(servicePort, kurtosisServersPortProtocol)
+				if err != nil {
+					return nil, nil, stacktrace.Propagate(err, "Expected to be able to create a port spec describing a public grpc proxy port from Kubernetes service port '%v', instead a non nil error was returned", servicePortName)
+				}
+				publicGrpcProxyPort = publicGrpcProxyPortSpec
+			}
+		}
+	}
+
+	if publicGrpcPort == nil || publicGrpcProxyPort == nil {
+		return nil, nil, stacktrace.NewError("Expected to get public port specs from Kubernetes service ports, instead got a nil pointer")
+	}
+
+	return publicGrpcPort, publicGrpcProxyPort, nil
+}
+
+func (backend *KubernetesKurtosisBackend) getAllEnclaveNamespaces(ctx context.Context) ([]apiv1.Namespace, error) {
+
+	matchLabels := getEnclaveMatchLabels()
+
+	namespaces, err := backend.kubernetesManager.GetNamespacesByLabels(ctx, matchLabels)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting the enclave namespace using labels '%+v'", matchLabels)
+	}
+
+	return namespaces.Items, nil
+}
+
+func (backend *KubernetesKurtosisBackend) getEnclaveNamespace(ctx context.Context, enclaveId enclave.EnclaveID) (*apiv1.Namespace, error) {
+
+	matchLabels := getEnclaveMatchLabels()
+	matchLabels[label_key_consts.EnclaveIDLabelKey.GetString()] = string(enclaveId)
+
+	namespaces, err := backend.kubernetesManager.GetNamespacesByLabels(ctx, matchLabels)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting the enclave namespace using labels '%+v'", matchLabels)
+	}
+
+	numOfNamespaces := len(namespaces.Items)
+	if numOfNamespaces == 0 {
+		return nil, stacktrace.NewError("No namespace matching labels '%+v' was found", matchLabels)
+	}
+	if numOfNamespaces > 1 {
+		return nil, stacktrace.NewError("Expected to find only one api container namespace for api container in enclave ID '%v', but '%v' was found; this is a bug in Kurtosis", enclaveId, numOfNamespaces)
+	}
+
+	resultNamespace := &namespaces.Items[0]
+
+	return resultNamespace, nil
+}
+
+func (backend *KubernetesKurtosisBackend) getEnclaveDataPersistentVolumeClaim(ctx context.Context, enclaveNamespaceName string, enclaveId enclave.EnclaveID) (*apiv1.PersistentVolumeClaim, error) {
+	matchLabels := getEnclaveMatchLabels()
+	matchLabels[label_key_consts.VolumeTypeLabelKey.GetString()] = label_value_consts.EnclaveDataVolumeTypeLabelValue.GetString()
+	matchLabels[label_key_consts.EnclaveIDLabelKey.GetString()] = string(enclaveId)
+
+	persistentVolumeClaims, err := backend.kubernetesManager.GetPersistentVolumeClaimsByLabels(ctx, enclaveNamespaceName, matchLabels)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting the enclave persistent volume claim using labels '%+v'", matchLabels)
+	}
+
+	numOfPersistentVolumeClaims := len(persistentVolumeClaims.Items)
+	if numOfPersistentVolumeClaims == 0 {
+		return nil, stacktrace.NewError("No persistent volume claim matching labels '%+v' was found", matchLabels)
+	}
+	if numOfPersistentVolumeClaims > 1 {
+		return nil, stacktrace.NewError("Expected to find only one enclave data persistent volume claim for enclave ID '%v', but '%v' was found; this is a bug in Kurtosis", enclaveId, numOfPersistentVolumeClaims)
+	}
+
+	resultPersistentVolumeClaim := &persistentVolumeClaims.Items[0]
+
+	return resultPersistentVolumeClaim, nil
+}
+
+func getEnclaveMatchLabels() map[string]string {
+	matchLabels := map[string]string{
+		label_key_consts.AppIDLabelKey.GetString():                label_value_consts.AppIDLabelValue.GetString(),
+		label_key_consts.KurtosisResourceTypeLabelKey.GetString(): label_value_consts.EnclaveKurtosisResourceTypeLabelValue.GetString(),
+	}
+	return matchLabels
 }