+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_manager/consts/kubernetes_consts.go
@@ -18,6 +18,7 @@ const (
 	PersistentVolumeClaimsKubernetesResource = "persistentvolumeclaims"
 
 	ClusterRoleKubernetesResourceType = "ClusterRole"
+	RoleKubernetesResourceType = "Role"
 
 	RbacAuthorizationApiGroup = "rbac.authorization.k8s.io"
 )