+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,8 @@
 # TBD
 ### Features
+* Added `KubernetesKurtosisBackend.GetEnclaves` functionality to kubernetes backend
+* Added `KubernetesKurtosisBackend.CreateAPIContainer`, `KubernetesKurtosisBackend.GetAPIContainers`, `KubernetesKurtosisBackend.StopAPIContainers` and `KubernetesKurtosisBackend.DestroyAPIContainers` 
+* Added `KubernetesKurtosisBackend.isPodRunningDeterminer` utility variable that we use for determine if a pod is running
 * Added `GetInClusterKubernetesKurtosisBackend` Kurtosis backend factory method to be used for pods inside Kubernetes cluster
 * `Network` objects returned by `DockerManager` will have the gateway IP and the IPs of the containers on the network